add scenario refresh button
