+++ app/packages/core/src/plugins/SchemaIO/components/NativeModelEvaluationView/evaluation/scenario/Actions.tsx
@@ -1,0 +1,49 @@
+import { Delete, Edit } from "@mui/icons-material";
+import MoreVertIcon from "@mui/icons-material/MoreVert";
+import {
+  Box,
+  IconButton,
+  ListItemIcon,
+  Menu,
+  MenuItem,
+  Typography,
+} from "@mui/material";
+import React from "react";
+
+export default function Actions(props) {
+  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+  const open = Boolean(anchorEl);
+  const handleClick = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorEl(event.currentTarget);
+  };
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
+
+  return (
+    <Box>
+      <IconButton onClick={handleClick}>
+        <MoreVertIcon />
+      </IconButton>
+      <Menu
+        id="actions-menu"
+        anchorEl={anchorEl}
+        open={open}
+        onClose={handleClose}
+      >
+        <MenuItem onClick={handleClose}>
+          <ListItemIcon>
+            <Edit fontSize="small" color="secondary" />
+          </ListItemIcon>
+          <Typography color="secondary">Edit</Typography>
+        </MenuItem>
+        <MenuItem onClick={handleClose}>
+          <ListItemIcon>
+            <Delete fontSize="small" color="error" />
+          </ListItemIcon>
+          <Typography color="error">Delete</Typography>
+        </MenuItem>
+      </Menu>
+    </Box>
+  );
+}
